+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Update list of known properties (used by PropertySpelling lint)
 * Fix bug where SassScript selectors referring to the current selector would
   result in a crash
+* Enhance `TrailingSemicolon` to check for more than one semicolon at the end
+  of a statement
 
 ## 0.28.0